Fix AppState singleton, update project config, and resolve share context import
